chore: implement subscriber page element
